+++ source/plugins/language.visual.wyapp/visual/definitions_date_time.js
@@ -28,6 +28,22 @@ module.exports = function (blockly) {
 		}
 	};
 
+	Blockly.Blocks['delay'] = {
+		init: function() {
+			this.setHelpUrl('https://projects.wyliodrin.com/wiki/languages/visual#delay');
+			this.setColour(180);
+			this.appendValueInput('millis')
+				.setCheck('Number')
+				.appendField('delay');
+			this.appendDummyInput()
+				.appendField(new Blockly.FieldDropdown([['milliseconds', '0'], ['microseconds', '1'], ['seconds', '2']]), 'type');
+			this.setInputsInline(true);
+			this.setPreviousStatement(true);
+			this.setNextStatement(true);
+			this.setTooltip('Wait for some specified period.');
+		}
+	};
+
 	Blockly.Blocks['format_numbers'] = {
 		init: function () {
 			this.setHelpUrl('http://www.example.com/');