+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
   * **added** social (email, facebook twitter and twilio) visual programming blocks
   * **added** Adafruit DHT sensor visual programming blocks for Raspberry Pi and BeagleBone Black
+  * **added** delay block to visual programming blocks
   * **feature** Wyapp PackageManager search with any case
   * **fixed** Wyapp FileManager dispalying double names
   * **fixed** (partially) scroll bug in Firefox (temporary fix)